Services updates + order customer service  + first controller

+++ src/main/java/com/PersonalProject/Jemo/services/Implemenation/ProductServiceImpl.java
@@ -1,10 +1,15 @@
 package com.PersonalProject.Jemo.services.Implemenation;
 
+import com.PersonalProject.Jemo.dto.ItemOrderCustomerDto;
+import com.PersonalProject.Jemo.dto.ItemOrderSupplierDto;
 import com.PersonalProject.Jemo.dto.ProductDto;
 import com.PersonalProject.Jemo.exception.EntityNotFoundException;
 import com.PersonalProject.Jemo.exception.EntityNotValidException;
 import com.PersonalProject.Jemo.exception.ErrorCodes;
-import com.PersonalProject.Jemo.repository.ProductRepository;
+import com.PersonalProject.Jemo.exception.OperationNotValidException;
+import com.PersonalProject.Jemo.model.ItemOrderCustomer;
+import com.PersonalProject.Jemo.model.ItemOrderSupplier;
+import com.PersonalProject.Jemo.repository.*;
 import com.PersonalProject.Jemo.services.ProductService;
 import com.PersonalProject.Jemo.validator.ProductValidator;
 import lombok.extern.slf4j.Slf4j;
@@ -20,10 +25,16 @@ import java.util.stream.Collectors;
 @Slf4j
 public class ProductServiceImpl implements ProductService {
     private  ProductRepository productRepository;
+    private ItemOrderCustomerRepository itemOrderCustomerRepository;
+    private ItemOrderSupplierRepository itemOrderSupplierRepository;
+
+
     @Autowired
-    public ProductServiceImpl(ProductRepository productRepository) {
+    public ProductServiceImpl(ProductRepository productRepository, ItemOrderCustomerRepository itemOrderCustomerRepository,ItemOrderSupplierRepository itemOrderSupplierRepository) {
         super();
         this.productRepository = productRepository;
+        this.itemOrderCustomerRepository = itemOrderCustomerRepository;
+        this.itemOrderSupplierRepository = itemOrderSupplierRepository;
 
     }
 
@@ -56,7 +67,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public ProductDto findById(Integer id) {
+    public ProductDto findById(Long id) {
         if (id == null){
             log.error("Product ID is Null");
             return null;
@@ -76,23 +87,8 @@ public class ProductServiceImpl implements ProductService {
                 .collect(Collectors.toList());
     }
 
-    /*@Override
-    public ProductDto update(ProductDto productDto, Integer id) {
-        List<String> errors = ProductValidator.validator(productDto);
-        if(!errors.isEmpty()){
-            log.error("Product not valid");
-            throw new EntityNotValidException("product invalid", ErrorCodes.PRODUCT_NOT_VALID, errors);
-        }
-        if (id == null){
-            log.error("Product ID is Null");
-             return null;
-        }
-
-        return productRepository.s;
-    }*/
-
     @Override
-    public List<ProductDto> findAllByCategoryId(Integer idCategory) {
+    public List<ProductDto> findAllByCategoryId(Long idCategory) {
         if (idCategory == null){
             log.error("Category ID is Null");
             return null;
@@ -102,11 +98,32 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public void delete(Integer id) {
+    public List<ItemOrderCustomerDto> findHistoryOrderClient(Long id) {
+        return itemOrderCustomerRepository.findAllByProductId(id).stream()
+                .map(ItemOrderCustomerDto::fromEntity).collect(Collectors.toList());
+    }
+
+    @Override
+    public List<ItemOrderSupplierDto> findHistoryOrderSupplier(Long id) {
+        return itemOrderSupplierRepository.findAllByProductId(id).stream()
+                .map(ItemOrderSupplierDto::fromEntity).collect(Collectors.toList());
+    }
+
+    @Override
+    public void delete(Long id) {
         if (id == null) {
             log.error("Product ID is Null");
-        } else {
-            productRepository.deleteById(id);
+            return;
         }
+        List<ItemOrderCustomer> itemOrderCustomers = itemOrderCustomerRepository.findAllByProductId(id);
+        if (!itemOrderCustomers.isEmpty()){
+            throw new OperationNotValidException("Product already in use(Order)",ErrorCodes.PRODUCT_ALREADY_IN_USE);
+        }
+        List<ItemOrderSupplier> itemOrderSuppliers = itemOrderSupplierRepository.findAllByProductId(id);
+        if (!itemOrderSuppliers.isEmpty()){
+            throw new OperationNotValidException("Product already in use(Supplier)",ErrorCodes.PRODUCT_ALREADY_IN_USE);
+        }
+            productRepository.deleteById(id);
+
     }
 }
